+++ src/components/EchartsEditor/EchartsEditor.test.tsx
@@ -1,16 +1,17 @@
-import React from 'react';
 import { getTemplateSrv } from '@grafana/runtime';
 import { CodeEditor, CodeEditorSuggestionItemKind } from '@grafana/ui';
 import { render, screen } from '@testing-library/react';
+import React from 'react';
+
 import {
-  CodeEditorSuggestions,
+  CODE_EDITOR_SUGGESTIONS,
   CodeLanguage,
   Editor,
   Format,
-  TestIds,
-  VisualCodeEditorSuggestions,
+  TEST_IDS,
+  VISUAL_CODE_EDITOR_SUGGESTIONS,
 } from '../../constants';
-import { EChartsEditor } from './EChartsEditor';
+import { EchartsEditor } from './EchartsEditor';
 
 /**
  * Mock @grafana/ui
@@ -56,7 +57,7 @@ describe('Echarts Editor', () => {
   };
 
   const getComponent = ({ ...restProps }: any, context = getContext()) => {
-    return <EChartsEditor item={{}} {...restProps} context={context} />;
+    return <EchartsEditor item={{}} {...restProps} context={context} />;
   };
 
   beforeEach(() => {
@@ -65,7 +66,7 @@ describe('Echarts Editor', () => {
 
   it('Should find component', async () => {
     render(getComponent({}));
-    expect(screen.getByTestId(TestIds.editor.root)).toBeInTheDocument();
+    expect(screen.getByTestId(TEST_IDS.editor.root)).toBeInTheDocument();
   });
 
   it('Should show mini map if value more than 100 symbols', () => {
@@ -163,7 +164,7 @@ describe('Echarts Editor', () => {
 
     render(getComponent({}));
 
-    expect(suggestionsResult).toEqual(expect.arrayContaining(CodeEditorSuggestions));
+    expect(suggestionsResult).toEqual(expect.arrayContaining(CODE_EDITOR_SUGGESTIONS));
     expect(suggestionsResult).toEqual(
       expect.arrayContaining([
         {
@@ -219,7 +220,7 @@ describe('Echarts Editor', () => {
       )
     );
 
-    expect(suggestionsResult).toEqual(expect.arrayContaining(VisualCodeEditorSuggestions));
+    expect(suggestionsResult).toEqual(expect.arrayContaining(VISUAL_CODE_EDITOR_SUGGESTIONS));
   });
 
   it('Should use JSON language for themeConfig item', () => {