+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export * from './EChartsEditor';
-export * from './EChartsPanel';
+export * from './EchartsEditor';
+export * from './EchartsPanel';
 export * from './VisualEditor';