+++ src/constants/echarts.ts
@@ -17,7 +17,7 @@ export enum Theme {
 /**
  * Theme options
  */
-export const ThemeOptions = [
+export const THEME_OPTIONS = [
   { value: Theme.DEFAULT, label: 'Default', description: 'Default ECharts theme' },
   { value: Theme.CUSTOM, label: 'Custom', description: 'Ability to use custom ECharts theme' },
 ];
@@ -25,7 +25,7 @@ export const ThemeOptions = [
 /**
  * Renderer Options
  */
-export const RendererOptions = [
+export const RENDERER_OPTIONS = [
   { value: Renderer.CANVAS, label: 'Canvas', description: 'More suitable for charts with a large number of elements.' },
   { value: Renderer.SVG, label: 'SVG', description: 'Has less memory usage, no blur when using the browser zoom.' },
 ];