+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Updated README and documentation (#214)
 - Add visual editor for working with data sources (#211)
+- Update ESLint configuration and refactoring (#237)
 
 ## 5.1.0 (2023-08-11)
 
@@ -123,7 +124,7 @@
 - Update to Grafana 9.3.1 (#97)
 - Add Google map support (#98)
 - Refactor Maps support (#99)
-- Update to Apache ECharts 5.4.1 (#101) 
+- Update to Apache ECharts 5.4.1 (#101)
 
 ### Bug fixes